Create project structure
