+++ FlowersApp/Features/Orders/ViewModels/OrdersViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  OrdersViewModel.swift
+//  FlowersApp
+//
+//  Created by Emil Micu on 24/04/2018.
+//  Copyright © 2018 Emil Micu. All rights reserved.
+//
+
+import Foundation
+
+
+final class OrdersViewModel: OrdersViewModelType {
+    
+    
+}
