+++ FlowersApp/Features/Orders/ViewModels/OrdersViewModelType.swift
@@ -1,0 +1,14 @@
+//
+//  OrdersViewModelType.swift
+//  FlowersApp
+//
+//  Created by Emil Micu on 24/04/2018.
+//  Copyright © 2018 Emil Micu. All rights reserved.
+//
+
+import Foundation
+
+protocol OrdersViewModelType: class {
+    
+    
+}
