+++ FlowersApp/Models/Order.swift
@@ -1,0 +1,17 @@
+//
+//  Order.swift
+//  FlowersApp
+//
+//  Created by Emil Micu on 24/04/2018.
+//  Copyright © 2018 Emil Micu. All rights reserved.
+//
+
+import Foundation
+
+
+struct Order {
+    var ID: Int?
+    var description: String?
+    var price: Double?
+    var deliverTo: String?
+}
