+++ FlowersApp/Networking/APIClient.swift
@@ -1,0 +1,14 @@
+//
+//  APIClient.swift
+//  FlowersApp
+//
+//  Created by Emil Micu on 24/04/2018.
+//  Copyright © 2018 Emil Micu. All rights reserved.
+//
+
+import Foundation
+
+final class APIClient {
+    
+    
+}
